Try catch dans functions.ps1 pour save-kudufolder2url.ps1

+++ scripts/Save-kudufolder2Url.ps1
@@ -13,21 +13,23 @@ param(
     [string] 
     $SlotName = "",
     [string] 
-    [ValidateSet("x","y","z")] 
+    [ValidateSet("x", "y", "z")] 
     $ZipFolder = "Z"
 )
 
 # Declaration de variables:
 if ($env:COMPUTERNAME -like "srvtfs01") { . "$DefaultWorkingDirectory\DevOps\scripts\Functions.ps1" }  else { . C:\templates\DevOps\scripts\Functions.ps1 }
-if ($webApp -like "Gumsolr") { $kuduPath = "server/solr/index/" } Else {$kuduPath = "App_Data/Logs/"} 
-if (!( Get-PSDrive -name z -ErrorAction SilentlyContinue)) {new-azdrive}
+if ($webApp -like "Gumsolr") { $kuduPath = "server/solr/index/" } Else { $kuduPath = "App_Data/Logs/" } 
+if (!( Get-PSDrive -name z -ErrorAction SilentlyContinue)) { new-azdrive }
 $WebappName = "$webapp-$Environnement"
 $resourceGroupName = (get-azureRmwebapp -name $webAppName ).resourcegroup
-$ZipFilePath = ("c:\temp\" + $environnement + "\" + $webAppName + (get-date -Format "yyyy-MM-dd") +  ".zip") 
-$Destination= ($ZipFolder + ":\" + $environnement + "\" + $webAppName + (get-date -Format "yyyy-MM-dd") +  ".zip") 
+$ZipFilePath = ("c:\temp\" + $environnement + "\" + $webAppName + (get-date -Format "yyyy-MM-dd") + ".zip") 
+$Destination = ($ZipFolder + ":\" + $environnement + "\" + $webAppName + (get-date -Format "yyyy-MM-dd") + ".zip") 
 
-Compress-KuduFolderToZipFile -Environnement $environnement -resourceGroupName $resourceGroupName -webAppName $webAppName -ZipFilePath $ZipFilePath -kuduPath $kuduPath 
-copy-item -Path $ZipFilePath  -Destination $Destination
-Get-ChildItem ($ZipFolder + ":\" + $environnement) | Where-Object {$_.LastWriteTime -lt (get-date).adddays(-10)} | remove-item
-remove-psdrive -name $ZipFolder
-Remove-Item -path $ZipFilePath -Confirm:$false
+$scrap = Compress-KuduFolderToZipFile -Environnement $environnement -resourceGroupName $resourceGroupName -webAppName $webAppName -ZipFilePath $ZipFilePath -kuduPath $kuduPath
+if ($scrap -notlike "false") {
+    copy-item -Path $ZipFilePath  -Destination $Destination
+    Get-ChildItem ($ZipFolder + ":\" + $environnement) | Where-Object { $_.LastWriteTime -lt (get-date).adddays(-10) } | remove-item
+    remove-psdrive -name $ZipFolder
+    Remove-Item -path $ZipFilePath -Confirm:$false
+} else { "Le folder $kuduPath n'existe pas!"}
